+++ include/grinch/fs/devfs.h
@@ -37,7 +37,7 @@ struct devfs_ops {
 			char *ubuf, size_t count);
 	ssize_t (*write)(struct devfs_node *node, struct file_handle *fh,
 			 const char *ubuf, size_t count);
-	long (*ioctl)(struct devfs_node *node, unsigned long op,
+	long (*ioctl)(struct devfs_node *node, unsigned int op,
 		      unsigned long arg);
 };
